Move main to version 6.0
* Upgrade to the latest dependencies
* Migrate samples to `jakarta` namespace
* Rework `cafe-scripted` to RSocket instead of already removed RMI
* Migrate JMS samples to Apache Artemis for Jakarta EE 9
* Disables mails tests since there is a requirement to migrate to Greenmail for Jakarta EE 9 support
* Rework `generatePom` task to a new `maven-publish` plugin style

+++ applications/cafe-scripted/src/main/java/org/springframework/integration/samples/cafe/demo/ControlBusMain.java
@@ -48,7 +48,7 @@ public class ControlBusMain {
 					logger.error("Interrupted", e);
 				}
 
-				totalDeliveries = (Integer)waiterMonitor.sendControlScript("waiter.totalDeliveries");
+				totalDeliveries = waiterMonitor.sendControlScript("waiter.totalDeliveries");
 
 				logger.info("Total cafe deliveries: " + totalDeliveries);
